fix change from last pull request
