+++ novapi/Team1/test_manual/novapi_new_program.py
@@ -1,16 +1,25 @@
-import novapi
-from mbuild import gamepad
-from mbuild import power_expand_board
-from mbuild.dual_rgb_sensor import dual_rgb_sensor_class
-from mbuild.smartservo import smartservo_class
 from mbuild.encoder_motor import encoder_motor_class
+from mbuild.smartservo import smartservo_class
+from mbuild.dual_rgb_sensor import dual_rgb_sensor_class
+from mbuild import power_expand_board
+from mbuild import gamepad
+import novapi
+import time
 
 # initialize the variables
 auto_stage = 0
 shoot = 0
 invert = 0
 feeddc = 1
-lrmode = 0 # Differentiate between shoot and arm control mode
+lrmode = 0  # Differentiate between shoot and arm control mode
+bp = 50
+
+# DC motors
+dc1_variable = "DC1"
+dc2_variable = "DC2"
+dc3_variable = "DC3"
+dc4_variable = "DC4"
+dc5_variable = "DC5"
 
 # Sensors
 dual_rgb_sensor_1 = dual_rgb_sensor_class("PORT2", "INDEX1")
@@ -39,13 +48,14 @@ def Manual():
     global invert
     global lrmode
     global feeddc
+    global bp
     LoadMe()
     while True:
         time.sleep(0.001)
         JoyRes.MovingJoystick(invert)
         ManualRes.InvertLED(invert)
         ManualRes.ControlLED(lrmode)
-        JoyRes.MultiControl(lrmode)
+        JoyRes.MultiControl(lrmode, bp)
 
         if gamepad.is_key_pressed("Up"):
             ManualRes.MoveForward()
@@ -60,9 +70,21 @@ def Manual():
             ManualRes.MoveRight()
 
         if gamepad.is_key_pressed("N1"):
-            smartservo_arm.move(10, 10)
+            power_expand_board.set_power("DC5", -20)
+        elif gamepad.is_key_pressed("N4"):
+            power_expand_board.set_power("DC5", 20)
+        else:
+            power_expand_board.stop("DC5")
 
         if gamepad.is_key_pressed("N2"):
+            if bp == 50:
+                bp = 100
+            elif bp == 100:
+                bp = 0
+            else:
+                bp = 50
+            while not not gamepad.is_key_pressed("N2"):
+                pass
             pass
 
         if gamepad.is_key_pressed("N3"):
@@ -73,9 +95,6 @@ def Manual():
             while not not gamepad.is_key_pressed("N3"):
                 pass
             pass
-
-        if gamepad.is_key_pressed("N4"):
-            smartservo_arm.move(-10, 10)
 
         # Switch shooting to arm control
         if gamepad.is_key_pressed("R2"):
@@ -96,9 +115,9 @@ def Manual():
 
         # Dc feed
         if feeddc == 1:
-            power_expand_board.set_power("DC2", -100)
+            power_expand_board.set_power(dc2_variable, -100)
         else:
-            power_expand_board.stop("DC2")
+            power_expand_board.stop(dc2_variable)
 
 
 def LoadMe():
@@ -106,10 +125,10 @@ def LoadMe():
 
     smartservo_arm.set_power(50)
     smartservo_arm.move(30, 10)
-    power_expand_board.set_power("BL1", 50)
-    power_expand_board.set_power("DC1", 50)
-    power_expand_board.stop("BL1")
-    power_expand_board.stop("DC1")
+    # power_expand_board.set_power("BL1", 50)
+    # power_expand_board.set_power(dc1_variable, 50)
+    # power_expand_board.stop("BL1")
+    # power_expand_board.stop(dc1_variable)
 
 
 class JoyRes:
@@ -129,34 +148,45 @@ class JoyRes:
 
         # Adjust LR slide tuning here
         if gamepad.get_joystick("Lx") != 0:
-
+            
             if gamepad.get_joystick("Lx") < 0:
-                Fl = Lx + 10
+                Fl = Lx - 20
+                Fr = Lx - 20
+
             if gamepad.get_joystick("Lx") > 10:
-                Fl = Lx - 10
+                Fl = Lx + 20
+                Fr = Lx + 20
+            
+            # Bring this back in case things wont go well
+            # Fr = Lx + Fr
+        # Fl = Lx
+        # Fr = Lx
+        Rl = Lx
+        Rr = Lx
+        # Encoder values. If the encoder motors config are changed even the
+        # slightest. change this one first then the inverted controls
 
-            Fr = Lx + Fr
-            Rl = Lx + Rl
-            Rr = Lx + Rr
-        # Encoder values. If the encoder motors config are changed even the slightest. change this one first then the inverted controls
-        EFl = 1 * (gamepad.get_joystick("Ly")
-                   + Fl - gamepad.get_joystick("Rx"))
-        EFr = -1 * (gamepad.get_joystick("Ly") - Fr
-                    + gamepad.get_joystick("Rx"))
-        ERl = 1 * (gamepad.get_joystick("Ly") - Rl
-                   - gamepad.get_joystick("Rx"))
-        ERr = -1 * (gamepad.get_joystick("Ly") + Rr
-                    + gamepad.get_joystick("Rx"))
+        vl = 0.8
 
-        if invert == 1:  # If the controls are inverted The arms are now the bot's front
-            ERr = 1 * (gamepad.get_joystick("Ly")
-                       - Fl - gamepad.get_joystick("Rx"))
-            ERl = -1 * (gamepad.get_joystick("Ly") + Fr
-                        + gamepad.get_joystick("Rx"))
-            EFr = 1 * (gamepad.get_joystick("Ly") + Rl
-                       - gamepad.get_joystick("Rx"))
-            EFl = -1 * (gamepad.get_joystick("Ly") - Rr
-                        + gamepad.get_joystick("Rx"))
+        EFl = vl * (gamepad.get_joystick("Ly") - Rl
+                    - gamepad.get_joystick("Rx"))
+        EFr = -vl * (gamepad.get_joystick("Ly") + Rr
+                     + gamepad.get_joystick("Rx"))
+        ERl = vl * (gamepad.get_joystick("Ly")
+                    + Fl - gamepad.get_joystick("Rx"))
+        ERr = -vl * (gamepad.get_joystick("Ly") - Fr
+                     + gamepad.get_joystick("Rx"))
+
+        if invert == 1:
+            # If the controls are inverted The arms are now the bot's front
+            EFr = vl * (gamepad.get_joystick("Ly")
+                        - Fl - gamepad.get_joystick("Rx"))
+            EFl = -vl * (gamepad.get_joystick("Ly") + Fr
+                         + gamepad.get_joystick("Rx"))
+            ERr = vl * (gamepad.get_joystick("Ly") + Rl
+                        - gamepad.get_joystick("Rx"))
+            ERl = -vl * (gamepad.get_joystick("Ly") - Rr
+                         + gamepad.get_joystick("Rx"))
         encoder_motor_M1.set_power(EFl)
         encoder_motor_M2.set_power(EFr)
         encoder_motor_M3.set_power(ERl)
@@ -165,34 +195,35 @@ class JoyRes:
     def TurretControl():
         global auto_stage
         # > 40 < 60
-        #if smartservo_updown.get_value("angle") > -43:
+        # if smartservo_updown.get_value("angle") > -43:
         #    smartservo_updown.move(-1, 10)
 
-        #if smartservo_updown.get_value("angle") < -60:
+        # if smartservo_updown.get_value("angle") < -60:
         #    smartservo_updown.move(1, 10)
-        #else:
+        # else:
         if smartservo_updown.get_value("angle") < -56:
             servo_value = smartservo_updown.get_value("angle")
             while servo_value < -56:
                 smartservo_updown.move(3, 10)
                 servo_value = smartservo_updown.get_value("angle")
 
-        smartservo_updown.move(-gamepad.get_joystick("Ry"), 10)
+        relative_angle = gamepad.get_joystick("Ry") / 3
+
+        smartservo_updown.move(relative_angle, 10)
 
     def FeedControl():
         if gamepad.is_key_pressed("L1"):
-            power_expand_board.set_power("DC1", -100)
+            power_expand_board.set_power(dc1_variable, -100)
         elif gamepad.is_key_pressed("L2"):
-            power_expand_board.set_power("DC1", 100)
+            power_expand_board.set_power(dc1_variable, 100)
         else:
             power_expand_board.stop("DC1")
-            0
 
     def ShootControl():
         if gamepad.is_key_pressed("R1"):
-            power_expand_board.set_power("DC3", -100)
+            power_expand_board.set_power(dc3_variable, -100)
         else:
-            power_expand_board.stop("DC3")
+            power_expand_board.stop(dc3_variable)
 
     def GrabControl():
         # DC4 L1 release R1 grab
@@ -206,18 +237,21 @@ class JoyRes:
         pass
 
     def HandControl():
-
-        smartservo_arm.move(gamepad.get_joystick("Ry"), 10)
+        power_expand_board.set_power("DC6", gamepad.get_joystick("Ry"))
+        # power_expand_board.set_power("DC7", gamepad.get_joystick("Ry"))
+        # smartservo_arm.move(gamepad.get_joystick("Ry"), 10)
         pass
 
-    def MultiControl(lc):
+    def MultiControl(lc, bp):
         if lc == 0:
             # Gun control mode
             JoyRes.TurretControl()
             JoyRes.ShootControl()
             JoyRes.FeedControl()
-            power_expand_board.set_power("BL1", 20)
-            power_expand_board.set_power("BL2", 20)
+
+            # < 15 -- 23 > : 25 max
+            power_expand_board.set_power("BL1", bp)
+            power_expand_board.set_power("BL2", bp)
         else:
             # Hand control mode
             JoyRes.HandControl()
